requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
black
isort
python-dotenv>=0.5.1

# market data
yfinance>=0.2.0
alpaca-py>=0.8.0
pandas>=1.3.0
numpy>=1.21.0

# visualization
plotly>=5.0.0

# deep learning
torch>=2.0.0
# Note: For CUDA support, install from pytorch.org: pip install torch --index-url https://download.pytorch.org/whl/cu118
# Note: For Apple Silicon (M1/M2/M3), the default pip install torch includes MPS support

# parallel processing
ray[tune]>=2.0.0
